requirements.txt
pytesseract
pdf2image
python-docx
pandas
scikit-learn
nltk
joblib